Move cache to store package
