+++ store/cache/cache_test.go
@@ -1,20 +1,21 @@
-package types_test
+package cache_test
 
 import (
 	"fmt"
 	"testing"
 
+	"github.com/cosmos/cosmos-sdk/store/cache"
+	iavlstore "github.com/cosmos/cosmos-sdk/store/iavl"
+	"github.com/cosmos/cosmos-sdk/store/types"
+
 	"github.com/stretchr/testify/require"
 	"github.com/tendermint/iavl"
 	dbm "github.com/tendermint/tm-db"
-
-	iavlstore "github.com/cosmos/cosmos-sdk/store/iavl"
-	"github.com/cosmos/cosmos-sdk/types"
 )
 
 func TestGetOrSetStoreCache(t *testing.T) {
 	db := dbm.NewMemDB()
-	mngr := types.NewCommitKVStoreCacheManager(types.DefaultCommitKVStoreCacheSize)
+	mngr := cache.NewCommitKVStoreCacheManager(cache.DefaultCommitKVStoreCacheSize)
 
 	sKey := types.NewKVStoreKey("test")
 	store := iavlstore.UnsafeNewStore(iavl.NewMutableTree(db, 100), 10, 10)
@@ -26,13 +27,13 @@ func TestGetOrSetStoreCache(t *testing.T) {
 
 func TestStoreCache(t *testing.T) {
 	db := dbm.NewMemDB()
-	mngr := types.NewCommitKVStoreCacheManager(types.DefaultCommitKVStoreCacheSize)
+	mngr := cache.NewCommitKVStoreCacheManager(cache.DefaultCommitKVStoreCacheSize)
 
 	sKey := types.NewKVStoreKey("test")
 	store := iavlstore.UnsafeNewStore(iavl.NewMutableTree(db, 100), 10, 10)
 	kvStore := mngr.GetStoreCache(sKey, store)
 
-	for i := uint(0); i < types.DefaultCommitKVStoreCacheSize*2; i++ {
+	for i := uint(0); i < cache.DefaultCommitKVStoreCacheSize*2; i++ {
 		key := []byte(fmt.Sprintf("key_%d", i))
 		value := []byte(fmt.Sprintf("value_%d", i))
 
@@ -45,7 +46,7 @@ func TestStoreCache(t *testing.T) {
 		require.Equal(t, res, value)
 	}
 
-	for i := uint(0); i < types.DefaultCommitKVStoreCacheSize*2; i++ {
+	for i := uint(0); i < cache.DefaultCommitKVStoreCacheSize*2; i++ {
 		key := []byte(fmt.Sprintf("key_%d", i))
 
 		kvStore.Delete(key)
